[IMP] hr_payroll: Move the BrowsableObject classes into a separated file
Purpose
=======

These classes are utils classes for the payslip computation.

It's better for the code readability to isolate them into a single file,
and keep the business logic into the payslip model code.
