+++ addons/hr_payroll/models/__init__.py
@@ -1,6 +1,7 @@
 #-*- coding:utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
+from . import browsable_object
 from . import hr_contract
 from . import hr_employee
 from . import res_config_settings
