+++ addons/hr_payroll/models/hr_payslip.py
@@ -8,6 +8,7 @@ from pytz import timezone
 
 from odoo import api, fields, models, tools, _
 from odoo.addons import decimal_precision as dp
+from odoo.addons.hr_payroll.models.browsable_object import BrowsableObject, InputLine, WorkedDays, Payslips
 from odoo.exceptions import UserError, ValidationError
 
 class HrPayslip(models.Model):
@@ -210,68 +211,12 @@ class HrPayslip(models.Model):
     # YTI TODO: pass recordset as argument
     @api.model
     def _get_payslip_lines(self, contract_ids, payslip_id):
+        # YTI TODO: Move into browsable_object.py ?
         def _sum_salary_rule_category(localdict, category, amount):
             if category.parent_id:
                 localdict = _sum_salary_rule_category(localdict, category.parent_id, amount)
             localdict['categories'].dict[category.code] = category.code in localdict['categories'].dict and localdict['categories'].dict[category.code] + amount or amount
             return localdict
-
-        class BrowsableObject(object):
-            def __init__(self, employee_id, dict, env):
-                self.employee_id = employee_id
-                self.dict = dict
-                self.env = env
-
-            def __getattr__(self, attr):
-                return attr in self.dict and self.dict.__getitem__(attr) or 0.0
-
-        class InputLine(BrowsableObject):
-            """a class that will be used into the python code, mainly for usability purposes"""
-            def sum(self, code, from_date, to_date=None):
-                if to_date is None:
-                    to_date = fields.Date.today()
-                self.env.cr.execute("""
-                    SELECT sum(amount) as sum
-                    FROM hr_payslip as hp, hr_payslip_input as pi
-                    WHERE hp.employee_id = %s AND hp.state = 'done'
-                    AND hp.date_from >= %s AND hp.date_to <= %s AND hp.id = pi.payslip_id AND pi.code = %s""",
-                    (self.employee_id, from_date, to_date, code))
-                return self.env.cr.fetchone()[0] or 0.0
-
-        class WorkedDays(BrowsableObject):
-            """a class that will be used into the python code, mainly for usability purposes"""
-            def _sum(self, code, from_date, to_date=None):
-                if to_date is None:
-                    to_date = fields.Date.today()
-                self.env.cr.execute("""
-                    SELECT sum(number_of_days) as number_of_days, sum(number_of_hours) as number_of_hours
-                    FROM hr_payslip as hp, hr_payslip_worked_days as pi
-                    WHERE hp.employee_id = %s AND hp.state = 'done'
-                    AND hp.date_from >= %s AND hp.date_to <= %s AND hp.id = pi.payslip_id AND pi.code = %s""",
-                    (self.employee_id, from_date, to_date, code))
-                return self.env.cr.fetchone()
-
-            def sum(self, code, from_date, to_date=None):
-                res = self._sum(code, from_date, to_date)
-                return res and res[0] or 0.0
-
-            def sum_hours(self, code, from_date, to_date=None):
-                res = self._sum(code, from_date, to_date)
-                return res and res[1] or 0.0
-
-        class Payslips(BrowsableObject):
-            """a class that will be used into the python code, mainly for usability purposes"""
-
-            def sum(self, code, from_date, to_date=None):
-                if to_date is None:
-                    to_date = fields.Date.today()
-                self.env.cr.execute("""SELECT sum(case when hp.credit_note = False then (pl.total) else (-pl.total) end)
-                            FROM hr_payslip as hp, hr_payslip_line as pl
-                            WHERE hp.employee_id = %s AND hp.state = 'done'
-                            AND hp.date_from >= %s AND hp.date_to <= %s AND hp.id = pl.slip_id AND pl.code = %s""",
-                            (self.employee_id, from_date, to_date, code))
-                res = self.env.cr.fetchone()
-                return res and res[0] or 0.0
 
         #we keep a dict with the result because a value can be overwritten by another rule with the same code
         result_dict = {}
